+++ relay/nakama/main.go
@@ -7,6 +7,7 @@ import (
 	"encoding/json"
 	"io"
 	"os"
+	"pkg.world.dev/world-engine/relay/nakama/events"
 	"strings"
 	"sync"
 
@@ -55,7 +56,7 @@ func InitModule(
 
 	initReceiptDispatcher(logger)
 
-	if err := initEventHub(ctx, logger, nk, globalCardinalAddress); err != nil {
+	if err := initEventHub(ctx, logger, nk, EventEndpoint, globalCardinalAddress); err != nil {
 		return eris.Wrap(err, "failed to init event hub")
 	}
 
@@ -104,9 +105,10 @@ func initEventHub(
 	ctx context.Context,
 	log runtime.Logger,
 	nk runtime.NakamaModule,
+	eventsEndpoint string,
 	cardinalAddress string,
 ) error {
-	eventHub, err := createEventHub(log, cardinalAddress)
+	eventHub, err := events.CreateEventHub(log, eventsEndpoint, cardinalAddress)
 	if err != nil {
 		return err
 	}
@@ -121,7 +123,7 @@ func initEventHub(
 	go func() {
 		channel := eventHub.Subscribe("main")
 		for event := range channel {
-			err := eris.Wrap(nk.NotificationSendAll(ctx, "event", map[string]interface{}{"message": event.message}, 1, true), "")
+			err := eris.Wrap(nk.NotificationSendAll(ctx, "event", map[string]interface{}{"message": event.Message}, 1, true), "")
 			if err != nil {
 				log.Error("error sending notifications: %s", eris.ToString(err, true))
 			}
